// src/app/shared/constants/url-enum.ts
export enum Url {
  LOGIN = '/login',
  MAIN = '/main',
  NEWPURCH = '/purchases/new',
  NEWINC = '/incomes/new',
  PURCH = '/purchases',
  INC = '/incomes',
  ALLPURCH = '/purchases/all',
  ALLINC = '/incomes/all',
  STATIST = '/statistics',
  GRAPHS = '/graphs',
  REGIST = '/register',
  AUTH = '/authenticate',
  UPD = '/update',
  BASE = 'http://localhost:5500'
}